refactor: move AI error handling logic to a dedicated module and update imports
